semana 3

+++ semana3/temperatura.c
@@ -2,11 +2,11 @@
 
 int main(int argc, char const *argv[])
 {
-  for (float i = 0; i <= 100; i++)
+  for (int i = 0; i <= 100; i++)
   {
     float f = (i * 9.0 / 5.0) + 32.0;
-    printf("%.2f C = %.2f F\n", i, f);
+    printf("%d C = %.2f F\n", i, f);
   }
-  
+
   return 0;
 }
